+++ RealmSwift/SyncSubscription.swift
@@ -452,7 +452,7 @@ extension SyncSubscriptionSet: Sequence {
     }
 }
 
-#if swift(>=5.6) && canImport(_Concurrency)
+#if canImport(_Concurrency)
 @available(macOS 10.15, tvOS 13.0, iOS 13.0, watchOS 6.0, *)
 extension SyncSubscriptionSet {
     /**