+++ build.sh
@@ -1268,11 +1268,11 @@ x.y.z Release notes (yyyy-MM-dd)
 <!-- ### Breaking Changes - ONLY INCLUDE FOR NEW MAJOR version -->
 
 ### Compatibility
-* Realm Studio: 11.0.0 or later.
+* Realm Studio: 11.0.0 - 12.0.0.
 * APIs are backwards compatible with all previous releases in the 10.x.y series.
-* Carthage release for Swift is built with Xcode 14.1.
+* Carthage release for Swift is built with Xcode 14.2.
 * CocoaPods: 1.10 or later.
-* Xcode: 13.1-14.1.
+* Xcode: 13.3-14.2.
 
 ### Internal
 * Upgraded realm-core from ? to ?