+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 x.y.z Release notes (yyyy-MM-dd)
 =============================================================
+
+Swift 5.5 is no longer supported. Swift 5.6 (Xcode 13.3) is now the minimum
+supported version.
+
+The prebuilt binary for Carthage is now build with Xcode 14.2.
+
 ### Enhancements
 * Improve performance of creating Projection objects and of change
   notifications on projections ([PR #8050](https://github.com/realm/realm-swift/pull/8050)).
@@ -35,11 +41,11 @@ x.y.z Release notes (yyyy-MM-dd)
 <!-- ### Breaking Changes - ONLY INCLUDE FOR NEW MAJOR version -->
 
 ### Compatibility
-* Realm Studio: 11.0.0 or later.
+* Realm Studio: 11.0.0 - 12.0.0. 13.0.0 is currently incompatible.
 * APIs are backwards compatible with all previous releases in the 10.x.y series.
-* Carthage release for Swift is built with Xcode 14.1.
+* Carthage release for Swift is built with Xcode 14.2.
 * CocoaPods: 1.10 or later.
-* Xcode: 13.1-14.1.
+* Xcode: 13.3-14.2.
 
 ### Internal
 * Upgraded realm-core from ? to ?